requirements.txt
streamlit
pandas
pycaret
python-dotenv
openai
instructor
pydantic
joblib
langfuse